+++ shopify/packages/ajax-cart/src/index.ts
@@ -107,6 +107,35 @@ const AjaxCart = mergeProps(Cart.query, {
   subscribe,
 });
 
+const oldFetch = window.fetch;
+
+// Override any mutation based fetch calls to the cart to ensure our cart state is always up to date.
+window.fetch = async function (...args: Parameters<typeof fetch>) {
+  let [input, options] = args;
+  let route: string;
+  if (input instanceof Request) route = input.url;
+  else if (input instanceof URL) route = input.toString();
+  else if (typeof input === "string") route = input;
+  else route = "";
+  if (!route.startsWith("/cart/")) return oldFetch(...args);
+
+  if (input instanceof Request && !options) {
+    options = {
+      headers: input.headers,
+    };
+  }
+
+  if (
+    options &&
+    options.headers &&
+    "x-sdk-variant" in options.headers &&
+    options.headers["x-sdk-variant"] === "brytdesigns"
+  )
+    return oldFetch(...args);
+  Cart.cancel();
+  return oldFetch(...args).finally(() => Cart.invalidate());
+};
+
 declare global {
   interface Window {
     BrytDesigns: {