+++ shopify/packages/ajax-cart-api/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @brytdesigns/shopify-ajax-cart-api
 
+## 1.0.9
+
+### Patch Changes
+
+- Intercept cart mutation calls to force update our cart
+
 ## 1.0.8
 
 ### Patch Changes